upstream(core): Move metrics push client to a separate crate
