+++ crates/iota-common/src/lib.rs
@@ -4,8 +4,6 @@
 
 pub mod backoff;
 pub mod logging;
-#[cfg(feature = "metrics")]
-pub mod metrics;
 pub mod random;
 pub mod random_util;
 pub mod stream_ext;
