+++ crates/iota-metrics-push-client/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "iota-metrics-push-client"
+version = "0.1.0"
+authors = ["IOTA Foundation <info@iota.org>"]
+edition = "2021"
+license = "Apache-2.0"
+publish = false
+
+[dependencies]
+# external dependencies
+anyhow.workspace = true
+fastcrypto.workspace = true
+prometheus.workspace = true
+reqwest.workspace = true
+serde.workspace = true
+snap.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+
+# internal dependencies
+iota-metrics.workspace = true
+iota-tls.workspace = true
+iota-types.workspace = true
