+++ crates/iota-node/Cargo.toml
@@ -40,7 +40,7 @@ url.workspace = true
 # internal dependencies
 bin-version.workspace = true
 iota-archival.workspace = true
-iota-common = { workspace = true, features = ["metrics"] }
+iota-common.workspace = true
 iota-config.workspace = true
 iota-core.workspace = true
 iota-grpc-server.workspace = true
@@ -49,6 +49,7 @@ iota-json-rpc.workspace = true
 iota-json-rpc-api.workspace = true
 iota-macros.workspace = true
 iota-metrics.workspace = true
+iota-metrics-push-client.workspace = true
 iota-names.workspace = true
 iota-network.workspace = true
 iota-network-stack.workspace = true
